+++ src/include/simeng/pipeline/PortAllocator.hh
@@ -33,7 +33,7 @@ class PortAllocator {
   /** Set function from DispatchIssueUnit to retrieve reservation
    * station sizes during execution. */
   virtual void setRSSizeGetter(
-      std::function<void(std::vector<uint64_t>&)> rsSizes) = 0;
+      std::function<void(std::vector<uint32_t>&)> rsSizes) = 0;
 
   /** Tick the port allocator to allow it to process internal tasks. */
   virtual void tick() = 0;
